# src/main/resources/openapi/Authentication.yml
openapi: '3.0.3'
info:
  title: Token Authentication API
  version: '0.2.1'
servers:
  - url: https://api.server.test/v1
...
paths:
...
  /api/auth/admin:
    get:
      tags:
        - 'admin'
      security:
        - bearerAuth: []
      summary: 'List of Users registrated'
      description: 'Access enabled only to Admin'
      operationId: getAllUsers
      parameters:
        - name: username
          in: query
          required: false
          description: 'Searches by part of email string'
          schema:
            type: string
        - name: role
          in: query
          required: false
          description: 'Searches by role'
          schema:
            $ref: '#/components/schemas/EnumRole'
      responses:
        '200':
          description: 'List of Users retreived successfuly!'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInfo'
        '403':
          description: 'Forbidden!'
    patch:
      tags:
        - 'admin'
      security:
        - bearerAuth: []
      summary: 'Alter User Role'
      description: 'Access enabled only to Admin'
      operationId: alterUserRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRoles'
      responses:
        '200':
          description: 'Patched User Roles successfully!'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileRoles'
        '403':
          description: 'Forbidden!'
        '404':
          description: 'User not found!'
components:
  schemas:
...
    Username:
      type: string
      format: email
      description: 'User e-mail for contact and to redefine lost password [UNIQUE]'
      example: 'example@example.com'
      maxLength: 50
...
    EnumRole:
      type: string
      enum:
        - 'ADMIN'
        - 'MANAGER'
        - 'USER'
      description: 'Client permissions. Must be hidden from Users'
...
    ProfileRoles:
      type: object
      required:
        - 'username'
        - 'roles'
      properties:
        username:
          $ref: '#/components/schemas/Username'
        roles:
          type: array
          uniqueItems: true
          minItems: 1
          items:
            $ref: '#/components/schemas/EnumRole'
...
    UserInfo:
      type: object
      required:
        - 'id'
        - 'username'
        - 'roles'
      properties:
        id:
          type: string
          format: uuid
        username:
          $ref: '#/components/schemas/Username'
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/EnumRole'
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
      bearerFormat: JWT